Fix type correctness: unify DL call fields types
This commit addresses type correctness issues by fixing root types instead
of adding casts, eliminating unnecessary type conversions throughout the
codebase.

Key changes:

1. Effect struct tick fields: Changed all `start` and `stop` fields in
   effect structs (cef_freeze, cef_bless, cef_potion, cef_curse, etc.)
   from `int` to `uint32_t` to properly represent tick values, which
   are inherently non-negative and can exceed INT_MAX.

2. DL struct call fields unification: Changed all `call_x1`, `call_y1`,
   `call_x2`, `call_y2`, `call_x3` fields in the DL struct from `int`
   to `int64_t` to unify their types. These fields store various
   parameters (coordinates, ticker values, colors, etc.) and should
   have consistent representation.

3. Function signature updates: Updated rendering function signatures
   to accept `int64_t` parameters directly:
   - render_display_strike, render_display_pulseback
   - render_draw_bless, render_draw_potion, render_draw_rain
   - render_draw_curve, draw_pixel, render_text_fmt
   Functions now convert internally where needed, removing casts at
   call sites.

4. Removed unnecessary casts: Eliminated all explicit casts in
   game_core.c's dl_play() function when calling rendering functions,
   as function signatures now match the stored types.

5. Time-related function updates:
   - dl_prefetch: changed parameter from `int` to `uint32_t`
   - sdl_pre_add: changed attick parameter from `int` to `uint32_t`

This approach fixes the root cause of type mismatches rather than
silencing warnings with casts, improving type safety and code clarity.

+++ src/astonia.h
@@ -162,7 +162,7 @@ DLL_EXPORT void paranoia(const char *format, ...) __attribute__((format(printf, 
 void display_messagebox(char *title, char *text);
 
 void *xmalloc(int size, int ID);
-void *xrealloc(void *ptr, int size, int ID);
+void *xrealloc(void *ptr, size_t size, int ID);
 void *xrecalloc(void *ptr, int size, int ID);
 void xfree(void *ptr);
 char *xstrdup(const char *src, int ID);
