+++ src/client/client.h
@@ -2,6 +2,7 @@
  * Part of Astonia Client (c) Daniel Brockhaus. Please read license.txt.
  */
 #include "./dll.h"
+#include <time.h>
 
 #define MAXCHARS 2048
 
@@ -185,7 +186,7 @@ struct player {
 };
 
 struct cef_generic {
-	int nr;
+	unsigned int nr;
 	int type;
 };
 
@@ -193,7 +194,7 @@ struct cef_shield {
 	int nr;
 	int type;
 	int cn;
-	int start;
+	uint32_t start;
 };
 
 struct cef_strike {
@@ -206,7 +207,7 @@ struct cef_strike {
 struct cef_ball {
 	int nr;
 	int type;
-	int start;
+	uint32_t start;
 	int frx, fry; // high precision coords
 	int tox, toy; // high precision coords
 };
@@ -214,7 +215,7 @@ struct cef_ball {
 struct cef_fireball {
 	int nr;
 	int type;
-	int start;
+	uint32_t start;
 	int frx, fry; // high precision coords
 	int tox, toy; // high precision coords
 };
@@ -222,7 +223,7 @@ struct cef_fireball {
 struct cef_edemonball {
 	int nr;
 	int type;
-	int start;
+	uint32_t start;
 	int base;
 	int frx, fry; // high precision coords
 	int tox, toy; // high precision coords
@@ -237,8 +238,8 @@ struct cef_flash {
 struct cef_explode {
 	int nr;
 	int type;
-	int start;
-	int base;
+	uint32_t start;
+	unsigned int base;
 };
 
 struct cef_warcry {
@@ -252,8 +253,8 @@ struct cef_bless {
 	int nr;
 	int type;
 	int cn;
-	int start;
-	int stop;
+	uint32_t start;
+	uint32_t stop;
 	int strength;
 };
 
@@ -261,15 +262,15 @@ struct cef_heal {
 	int nr;
 	int type;
 	int cn;
-	int start;
+	uint32_t start;
 };
 
 struct cef_freeze {
 	int nr;
 	int type;
 	int cn;
-	int start;
-	int stop;
+	uint32_t start;
+	uint32_t stop;
 };
 
 struct cef_burn {
@@ -282,13 +283,13 @@ struct cef_burn {
 struct cef_mist {
 	int nr;
 	int type;
-	int start;
+	uint32_t start;
 };
 
 struct cef_pulse {
 	int nr;
 	int type;
-	int start;
+	uint32_t start;
 };
 
 struct cef_pulseback {
@@ -302,8 +303,8 @@ struct cef_potion {
 	int nr;
 	int type;
 	int cn;
-	int start;
-	int stop;
+	uint32_t start;
+	uint32_t stop;
 	int strength;
 };
 
@@ -322,8 +323,8 @@ struct cef_curse {
 	int nr;
 	int type;
 	int cn;
-	int start;
-	int stop;
+	uint32_t start;
+	uint32_t stop;
 	int strength;
 };
 
@@ -343,7 +344,7 @@ struct cef_firering {
 	int nr;
 	int type;
 	int cn;
-	int start;
+	uint32_t start;
 };
 
 struct cef_bubble {
@@ -470,9 +471,9 @@ DLL_EXPORT extern int mana;
 DLL_EXPORT extern int rage;
 DLL_EXPORT extern int endurance;
 DLL_EXPORT extern int lifeshield;
-DLL_EXPORT extern int experience;
-DLL_EXPORT extern int experience_used;
-DLL_EXPORT extern int gold;
+DLL_EXPORT extern uint32_t experience;
+DLL_EXPORT extern uint32_t experience_used;
+DLL_EXPORT extern uint32_t gold;
 DLL_EXPORT extern uint32_t tick;
 extern int lasttick; // ticks in inbuf
 extern int q_size;
@@ -483,10 +484,10 @@ DLL_EXPORT extern unsigned int csprite; // and sprite
 DLL_EXPORT extern int con_cnt;
 DLL_EXPORT extern int con_type;
 DLL_EXPORT extern char con_name[80];
-DLL_EXPORT extern int container[CONTAINERSIZE];
-DLL_EXPORT extern int price[CONTAINERSIZE];
-DLL_EXPORT extern int itemprice[CONTAINERSIZE];
-DLL_EXPORT extern int cprice;
+DLL_EXPORT extern uint32_t container[CONTAINERSIZE];
+DLL_EXPORT extern uint32_t price[CONTAINERSIZE];
+DLL_EXPORT extern uint32_t itemprice[CONTAINERSIZE];
+DLL_EXPORT extern uint32_t cprice;
 DLL_EXPORT extern int protocol_version;
 
 extern int looklevel;
@@ -507,7 +508,7 @@ DLL_EXPORT extern uint16_t originy;
 extern int sockstate;
 extern time_t socktimeout;
 DLL_EXPORT extern char *target_server;
-extern int target_port;
+extern uint16_t target_port;
 extern int kicked_out;
 
 extern char **game_skilldesc;
@@ -517,15 +518,15 @@ DLL_EXPORT extern uint16_t act;
 DLL_EXPORT extern uint16_t actx;
 DLL_EXPORT extern uint16_t acty;
 
-DLL_EXPORT extern int lookinv[12];
-DLL_EXPORT extern int looksprite, lookc1, lookc2, lookc3;
+DLL_EXPORT extern uint32_t lookinv[12];
+DLL_EXPORT extern uint32_t looksprite, lookc1, lookc2, lookc3;
 DLL_EXPORT extern char look_name[80];
 DLL_EXPORT extern char look_desc[1024];
 
 DLL_EXPORT extern uint32_t realtime;
-DLL_EXPORT extern int mil_exp;
-extern int display_gfx;
-extern int display_time;
+DLL_EXPORT extern uint32_t mil_exp;
+extern unsigned int display_gfx;
+extern uint32_t display_time;
 
 DLL_EXPORT extern char pent_str[7][80];
 
@@ -535,7 +536,7 @@ DLL_EXPORT extern struct shrine_ppd shrine;
 void cmd_text(char *text);
 DLL_EXPORT int mapmn(int x, int y);
 int find_cn_ceffect(int cn, int skip);
-int find_ceffect(int fn);
+int find_ceffect(unsigned int fn);
 DLL_EXPORT int level2exp(int level);
 DLL_EXPORT int exp2level(int val);
 DLL_EXPORT int raise_cost(int v, int n);
@@ -559,7 +560,7 @@ void cmd_raise(int vn);
 void cmd_con(int pos);
 void cmd_look_con(int pos);
 void cmd_drop_gold(void);
-void cmd_take_gold(int amount);
+void cmd_take_gold(uint32_t amount);
 void cmd_junk_item(void);
 void cmd_getquestlog(void);
 void cmd_reopen_quest(int nr);
@@ -575,7 +576,7 @@ void cl_ticker(void);
 int close_client(void);
 int is_char_ceffect(int type);
 
-extern int server_cycles;
+extern double server_cycles;
 extern int change_area;
 extern int login_done;
 extern unsigned int unique;
